requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
autogen==1.0.16
autopep8==2.0.4
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
defusedxml==0.7.1
distro==1.9.0
dnspython==2.6.1
docopt==0.6.2
docutils==0.20.1
frozenlist==1.4.1
grpcio==1.62.1
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
importlib_metadata==7.0.2
iniconfig==2.0.0
isodate==0.6.1
jaraco.classes==3.3.1
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-path==0.3.2
jsonschema-specifications==2023.12.1
keyring==24.3.1
lazy-object-proxy==1.10.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
more-itertools==10.2.0
motor==3.3.2
multidict==6.0.5
nest-asyncio==1.6.0
nh3==0.2.15
numpy==1.26.4
openai==1.14.1
openapi-core==0.19.0
openapi-schema-validator==0.6.2
openapi-spec-validator==0.7.1
packaging==24.0
parse==1.20.1
pathable==0.4.3
pkginfo==1.10.0
pluggy==1.4.0
prance==23.6.21.0
pybars4==0.9.13
pycodestyle==2.11.1
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.17.2
PyMeta3==0.5.1
pymongo==4.6.2
pytest==8.1.1
python-dotenv==1.0.1
PyYAML==6.0.1
readme_renderer==43.0
referencing==0.31.1
regex==2023.12.25
requests==2.31.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986==2.0.0
rich==13.7.1
rpds-py==0.18.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
scipy==1.12.0
semantic-kernel==0.9.3b1
six==1.16.0
sniffio==1.3.1
tqdm==4.66.2
twine==5.0.0
typing_extensions==4.10.0
urllib3==2.2.1
Werkzeug==3.0.1
yarl==1.9.4
zipp==3.18.1
